duui-suprisal/requirements.txt
numpy==1.24.4
dkpro_cassis==0.9.1
fastapi==0.104.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
uvicorn==0.23.2
starlette==0.27.0
minicons
nltk
pandas
transformers
huggingface_hub
